requirements.txt
# 核心依赖
requests==2.31.0
python-dotenv==1.0.0

# 开发依赖
pytest==7.4.3
black==23.11.0
flake8==6.1.0